feat(Double): change Double to BigDecimal

+++ src/main/scala/com/rfna/checking_account/models/Balance.scala
@@ -1,5 +1,5 @@
 package com.rfna.checking_account.models
 
-case class Balance(value: Double) extends Ordered[Balance] {
+case class Balance(value: BigDecimal) extends Ordered[Balance] {
   override def compare(that: Balance) = this.value.compare(that.value)
 }
