+++ src/main/scala/com/rfna/checking_account/models/Debt.scala
@@ -2,7 +2,7 @@ package com.rfna.checking_account.models
 
 import java.time.LocalDate
 
-case class Debt(start: LocalDate, end: Option[LocalDate], amount: Double)
+case class Debt(start: LocalDate, end: Option[LocalDate], amount: BigDecimal)
 
 object Debt {
   def fromPeriod(start: Option[SimpleStatement], end: Option[SimpleStatement]): Debt = {
