+++ src/main/scala/com/rfna/checking_account/models/SimpleStatement.scala
@@ -2,4 +2,4 @@ package com.rfna.checking_account.models
 
 import java.time.LocalDate
 
-case class SimpleStatement(date: LocalDate, balance: Double)
+case class SimpleStatement(date: LocalDate, balance: BigDecimal)
